--- AdventureInnovation/AdventureInnovation/routes/web.php.old
@@ -1,18 +1,0 @@
-<?php
-
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
-
-Route::get('/', 'GuideController@khalid');
-
-// Route::get('/', function () {
-//     return view('welcome');
-// });
