--- AdventureInnovation/AdventureInnovation/routes/web.php.old3
@@ -1,36 +1,0 @@
-<?php
-
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
-
-//Route::get('/', 'GuideController@khalid');
-
-// Route::get('/', function () {
-//     return view('welcome');
-// });
-
-Route::get('/', function () {
-    return view('homepage');
-});
-
-Route::get('homepage', function () {
-    return view('homepage');
-});
-
-Route::get('contact', function()
-{
-    return view('contact');
-});
-
-Route::get('companyContact', function()
-{
-    return view('companyContact');
-});
